datastore.CanSendWithRateLimit
canSendWithRateLimit looks up the last time we sent a given (user profile + email template)
combination in the database, and returns whether we're past the given rateLimit duration.

For example, if we want to ensure we don't send `template_1` to `profile_1` more than once
per hour, we could use it like this:

> canSendWithRateLimit(profile_1, template_1, time.Duration(1) * time.Hour)

note that a rateLimit of `nil` means *unlimited* and will always return true. *use with care!*

for emails intended to be sent only once (e.g. onboarding emails), consider using 1 year.
this way, we avoid completely stale information, and the person will get an (infrequent) reminder
that we still hold their information, giving them an opportunity to ask us to delete it.