API: return team roster & signature unencrypted
(Note: continue to return it encrypted too for older versions of fk)

It's causing us trouble because we want users to be able to check if
their team has been updated without having to unlock their key (from
cron, for example)

Given that team rosters are identified by a difficult-to-guess UUID, and aren't
particularly sensitive, I think it's acceptable to return the roster unencrypted.

(In future I'd like to implement "soft" authentication in the form of an
API access token that we issue when a key is uploaded / updated. This
would protect against spoofing a key, which can currently be used to
gain some metadata.)