bump deps to support team.Version
`version` is a new field being added to the roster.

new versions of `fk` will send this field in the roster, so we must
support it without raising an error.

old versions will omit the field, in which case it'll default to 0. in
the short term, this is OK (it's equivalent to the current state)